mini assignment 2/21
Hi @nrdowling @hjennybo,

I finished the mini assignment for Tuesday and finally found a way to make the graph that Shannon wanted (Yay!). I am having trouble referring to the image I included within my text and have spent the last hour or so trying to mess around with the code. I know this assignment is already past due, and I appreciate the extension so I wanted to push the work I completed. Thanks for understanding!